mcp/image_agent/requirements.txt
mcp
requests>=2.32.3
httpx>=0.27.0
beautifulsoup4>=4.12.3
tenacity>=9.0.0
python-dotenv>=1.0.1
Pillow>=10.4.0
python-frontmatter>=1.1.0
PyYAML>=6.0.2
tqdm>=4.66.4
filetype>=1.2.0
aiofiles>=24.1.0
